cpufreq: Resolve CPUFREQ_NOTIFY issue
When the cpufreq_stat receives the frequency policy notification,
it checks if the frequency table and stats table are present for the cpu.
If both are present, it returns with -EBUSY error to the notification.
Due to this, none of the other registered modules receive this
notification.

If the stat table and sysfs nodes are already created, return 0 instead
of returning EBUSY error.

Change-Id: I1482b81acec92e8f0d1de75f7f67aef7f9094003
Signed-off-by: Archana Sathyakumar <asathyak@codeaurora.org>